Move specialized functions into specialized reconcilers
